LPS-143554 Extract interface
